move gearview creation to gears.ts
