+++ src/lib/components/App.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import { createGearView } from '$lib/scripts/gears';
 	import { loadStatConstants } from '$lib/scripts/json-loader';
 	import { loadObject, openImageDB } from '$lib/scripts/loader';
 	import { updateWeaponMatrix } from '$lib/scripts/loadout';
@@ -16,7 +17,6 @@
 	import { AppWindowIcon } from '@lucide/svelte';
 	import { onMount, type Component } from 'svelte';
 	import Dialog from './Dialog.svelte';
-	import { createGearView } from './nav/GearPage.svelte';
 	import ReadySignal from './ReadySignal.svelte';
 	import Toolbar from './Toolbar.svelte';
 
@@ -71,9 +71,7 @@
 
 		// processing
 		await loadStatConstants(); // need this for gear proecssing
-		Promise.all(
-			$user_gears.map((gear) => createGearView(gear, false, $user_loadouts, $current_loadout))
-		).then((gearViews) => {
+		Promise.all($user_gears.map((gear) => createGearView(gear, false))).then((gearViews) => {
 			$gear_views = gearViews;
 			console.log('Done processing user_gears');
 		});
