+++ src/lib/components/nav/GearPage.svelte
@@ -1,175 +1,3 @@
-<script module lang="ts">
-	export const RAINBOW_TITAN_STATS: TitanStat[] = [
-		'titan_flame_atk',
-		'titan_frost_atk',
-		'titan_volt_atk',
-		'titan_phys_atk'
-	];
-
-	export function getRollValue(stat: Stat, value: number): number {
-		const stc = STAT_CONSTANTS[stat];
-		return ((value - stc.base) * 2) / (stc.high_roll + stc.low_roll);
-	}
-
-	export function getTitanValue(stat: Stat, value: number): number {
-		const stc = STAT_CONSTANTS[stat];
-		return stc.titan_base + stc.titan_multiplier * (value - stc.base);
-	}
-
-	export function reverseTitanValue(stat: Stat, titanValue: number): number {
-		const stc = STAT_CONSTANTS[stat];
-		return (
-			(titanValue - stc.titan_base + stc.titan_multiplier * stc.base) / (1 + stc.titan_multiplier)
-		);
-	}
-
-	export async function createGearView(
-		gear: UserGear,
-		equip: boolean = false,
-		user_loadouts: AllLoadouts,
-		current_loadout: string
-	): Promise<GearView> {
-		const stats: GearViewStatLong[] = [];
-		const derived: GearViewStatShort[] = [];
-		let id: number = -1;
-		let part: GearPart = GearPart.UNKNOWN;
-		let hash = '';
-
-		Object.entries(gear).forEach(([key, value]) => {
-			switch (key) {
-				case 'id':
-					id = value as number;
-					break;
-				case 'part':
-					part = value as GearPart;
-					hash += value;
-					break;
-				case 'dateAdded':
-					break;
-				default:
-					const isTitan = key.startsWith('titan_');
-					const value_format = key.includes('_percent') ? 'float3d' : 'int';
-
-					let _stat: Stat;
-					let _stat_value: number;
-					let stat_label: string;
-					let stat_value_label: string;
-
-					let titan_key: TitanStat;
-					let titan_label: string;
-					let titan_value: number;
-					let titan_value_label: string;
-
-					if (isTitan) {
-						// derive normal stat from corresponding titan stat
-						titan_key = key as TitanStat;
-						titan_label = STAT_LABELS[titan_key] ?? key;
-						titan_value = Number(value);
-						titan_value_label = formatValue(value_format, value as string);
-
-						_stat = key.replace('titan_', '') as Stat;
-						_stat_value = reverseTitanValue(_stat, titan_value);
-						stat_label = STAT_LABELS[_stat] ?? key;
-						stat_value_label = formatValue(value_format, _stat_value.toString());
-					} else {
-						_stat = key as Stat;
-						_stat_value = Number(value);
-						stat_label = STAT_LABELS[_stat] ?? key;
-						stat_value_label = formatValue(value_format, _stat_value.toString());
-
-						titan_key = 'titan_' + _stat;
-						titan_label = 'Titan ' + stat_label;
-						titan_value = getTitanValue(_stat, _stat_value) + _stat_value;
-						titan_value_label = formatValue(value_format, titan_value.toString());
-					}
-
-					stats.push({
-						stat: _stat,
-						stat_label,
-						value: _stat_value,
-						value_label: stat_value_label,
-						roll: getRollValue(_stat, _stat_value),
-						titan_stat_label: titan_label,
-						titan_value_label: titan_value_label
-					});
-
-					// normal stat
-					derived.push({
-						stat: _stat,
-						stat_label,
-						value: _stat_value,
-						value_label: stat_value_label
-					});
-
-					// titan stat
-					derived.push({
-						stat: titan_key as TitanStat,
-						stat_label: titan_label,
-						value: titan_value,
-						value_label: titan_value_label
-					});
-
-					hash += _stat + stat_value_label;
-
-					break;
-			}
-		});
-		stats.sort((a, b) => (b.roll ?? 0) - (a.roll ?? 0));
-
-		// get highest roll stat
-		const bestRoll = stats[0];
-
-		if (bestRoll.stat.includes('_atk') && !bestRoll.stat.includes('percent')) {
-			const eleAtkStats = stats.filter(
-				(stat) => stat.stat.includes('_atk') && !stat.stat.includes('percent')
-			);
-
-			if (eleAtkStats.length >= 2) {
-				// rainbow gear!
-
-				const rainbowTitanValue =
-					getTitanValue(bestRoll.stat, bestRoll.value) * 0.95 + bestRoll.value;
-				const rainbowTitanValueLabel = formatValue('int', rainbowTitanValue.toString());
-				RAINBOW_TITAN_STATS.forEach((rainbowStat) => {
-					const statIdx = derived.findIndex((der) => der.stat === rainbowStat);
-
-					if (statIdx === -1) {
-						// add rainbow stat if missing
-						derived.push({
-							stat: rainbowStat,
-							stat_label: STAT_LABELS[rainbowStat],
-							value: rainbowTitanValue,
-							value_label: rainbowTitanValueLabel
-						});
-					} else {
-						// replace value if rainbow is higher
-						if (rainbowTitanValue > derived[statIdx].value) {
-							derived[statIdx].value = rainbowTitanValue;
-							derived[statIdx].value_label = rainbowTitanValueLabel;
-						}
-					}
-				});
-			}
-
-			// todo: the same for atk%
-		}
-
-		const isEquipped =
-			equip ||
-			(part !== GearPart.UNKNOWN &&
-				user_loadouts[current_loadout].equipped_gears[gear.part as ValidGearPart] === gear.id);
-
-		return {
-			id,
-			part,
-			stats,
-			hash,
-			derived,
-			isEquipped
-		};
-	}
-</script>
-
 <script lang="ts">
 	import ActionToolbar from '$lib/components/ActionToolbar.svelte';
 	import GearInfo from '$lib/components/dialog/GearInfo.svelte';
@@ -177,11 +5,14 @@
 	import UploadScreenshot from '$lib/components/dialog/UploadScreenshot.svelte';
 	import FlexGrid from '$lib/components/FlexGrid.svelte';
 	import StatIcon from '$lib/components/StatIcon.svelte';
-	import { ALL_STATS_REGEX, GearPart, VALID_GEAR_PARTS } from '$lib/scripts/gears.ts';
-	import { STAT_CONSTANTS } from '$lib/scripts/json-loader';
+	import {
+		ALL_STATS_REGEX,
+		createGearView,
+		GearPart,
+		VALID_GEAR_PARTS
+	} from '$lib/scripts/gears.ts';
 	import { saveObject } from '$lib/scripts/loader.ts';
 	import { ActionType } from '$lib/scripts/nav-metadata';
-	import { STAT_LABELS } from '$lib/scripts/stats.ts';
 	import {
 		current_loadout,
 		gear_views,
@@ -189,15 +20,11 @@
 		user_gears,
 		user_loadouts
 	} from '$lib/scripts/stores';
-	import { formatValue } from '$lib/scripts/validation.ts';
 	import type {
-		AllLoadouts,
 		AllStats,
 		GearAugment,
 		GearSearchView,
 		GearView,
-		GearViewStatLong,
-		GearViewStatShort,
 		StatGearUser as Stat,
 		StatGearTitan as TitanStat,
 		UserGear,
@@ -352,154 +179,6 @@
 		}
 
 		saveObject('loadouts_v1', $user_loadouts);
-	}
-
-	async function createGearView(gear: UserGear, equip: boolean = false): Promise<GearView> {
-		const stats: GearViewStatLong[] = [];
-		const derived: GearViewStatShort[] = [];
-		let id: number = -1;
-		let part: GearPart = GearPart.UNKNOWN;
-		let hash = '';
-		let augment: [GearAugment, number] | undefined = undefined;
-
-		Object.entries(gear).forEach(([key, value]) => {
-			switch (key) {
-				case 'id':
-					id = value as number;
-					break;
-				case 'part':
-					part = value as GearPart;
-					hash += value;
-					break;
-				case 'dateAdded':
-					break;
-				case 'augment':
-					augment = value as [GearAugment, number];
-					break;
-				default:
-					const isTitan = key.startsWith('titan_');
-					const value_format = key.includes('_percent') ? 'float3d' : 'int';
-
-					let _stat: Stat;
-					let _stat_value: number;
-					let stat_label: string;
-					let stat_value_label: string;
-
-					let titan_key: TitanStat;
-					let titan_label: string;
-					let titan_value: number;
-					let titan_value_label: string;
-
-					if (isTitan) {
-						// derive normal stat from corresponding titan stat
-						titan_key = key as TitanStat;
-						titan_label = STAT_LABELS[titan_key] ?? key;
-						titan_value = Number(value);
-						titan_value_label = formatValue(value_format, value as string);
-
-						_stat = key.replace('titan_', '') as Stat;
-						_stat_value = reverseTitanValue(_stat, titan_value);
-						stat_label = STAT_LABELS[_stat] ?? key;
-						stat_value_label = formatValue(value_format, _stat_value.toString());
-					} else {
-						_stat = key as Stat;
-						_stat_value = Number(value);
-						stat_label = STAT_LABELS[_stat] ?? key;
-						stat_value_label = formatValue(value_format, _stat_value.toString());
-
-						titan_key = 'titan_' + _stat;
-						titan_label = 'Titan ' + stat_label;
-						titan_value = getTitanValue(_stat, _stat_value) + _stat_value;
-						titan_value_label = formatValue(value_format, titan_value.toString());
-					}
-
-					stats.push({
-						stat: _stat,
-						stat_label,
-						value: _stat_value,
-						value_label: stat_value_label,
-						roll: getRollValue(_stat, _stat_value),
-						titan_stat_label: titan_label,
-						titan_value_label: titan_value_label
-					});
-
-					// normal stat
-					derived.push({
-						stat: _stat,
-						stat_label,
-						value: _stat_value,
-						value_label: stat_value_label
-					});
-
-					// titan stat
-					derived.push({
-						stat: titan_key as TitanStat,
-						stat_label: titan_label,
-						value: titan_value,
-						value_label: titan_value_label
-					});
-
-					hash += _stat + stat_value_label;
-
-					break;
-			}
-		});
-		stats.sort((a, b) => (b.roll ?? 0) - (a.roll ?? 0));
-
-		// get highest roll stat
-		const bestRoll = stats[0];
-
-		if (bestRoll.stat.includes('_atk') && !bestRoll.stat.includes('percent')) {
-			const eleAtkStats = stats.filter(
-				(stat) => stat.stat.includes('_atk') && !stat.stat.includes('percent')
-			);
-
-			if (eleAtkStats.length >= 2) {
-				// rainbow gear!
-
-				const rainbowTitanValue =
-					getTitanValue(bestRoll.stat, bestRoll.value) * 0.95 + bestRoll.value;
-				const rainbowTitanValueLabel = formatValue('int', rainbowTitanValue.toString());
-				RAINBOW_TITAN_STATS.forEach((rainbowStat) => {
-					const statIdx = derived.findIndex((der) => der.stat === rainbowStat);
-
-					if (statIdx === -1) {
-						// add rainbow stat if missing
-						derived.push({
-							stat: rainbowStat,
-							stat_label: STAT_LABELS[rainbowStat],
-							value: rainbowTitanValue,
-							value_label: rainbowTitanValueLabel
-						});
-					} else {
-						// replace value if rainbow is higher
-						if (rainbowTitanValue > derived[statIdx].value) {
-							derived[statIdx].value = rainbowTitanValue;
-							derived[statIdx].value_label = rainbowTitanValueLabel;
-						}
-					}
-				});
-			}
-
-			// todo: the same for atk%
-		}
-
-		// TODO: add other derived stats
-
-		const isEquipped =
-			equip ||
-			(part !== GearPart.UNKNOWN &&
-				$user_loadouts[$current_loadout].equipped_gears[gear.part as ValidGearPart] === gear.id);
-
-		return {
-			id,
-			part,
-			stats,
-			hash,
-			derived,
-			isEquipped,
-			augment
-		};
 	}
 
 	function addNewGear(gear: UserGear, equip: boolean = false) {
